requirements.txt
autopep8==2.0.0
click==8.1.3
Flask==2.2.2
Flask-Login==0.6.2
Flask-MySQLdb==1.0.1
Flask-WTF==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
mysqlclient==2.1.1
protobuf==3.20.1
pycodestyle==2.9.1
tomli==2.0.1
Werkzeug==2.2.2
WTForms==3.0.1
